requirements.txt
attrs==20.1.0
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
six==1.15.0
toml==0.10.1
flask==1.1.2
PyYAML==5.3.1